rename queuesim module to queue
