+++ simulator.py
@@ -1,4 +1,4 @@
-from queuesim import Queue, Connection, EXTERIOR
+from queue import Queue, Connection, EXTERIOR
 from scheduler import Scheduler, Event, EventType
 from rand.linearcongruent import RandomGenerator
 import argparse
